Change word data type from `long` to `int`
Let's focus on 8-bit to 32-bit system.

One thing that I don't know what to do about remained, system call.
As system call is an essential component of operating system, I was
thinking making it built-in not configurable in `CONFIGURE`. But what
makes me hesitating is that it would be expensive to implement in 8-bit
AVR. I can't stop doubt system performance aside from technical issues.

+++ drivers/console.c
@@ -25,7 +25,7 @@ static size_t usart_read(int id, void *buf, size_t size)
 {
 	int data;
 	char *c = (char *)buf;
-	unsigned long irqflag;
+	unsigned int irqflag;
 
 	spin_lock_irqsave(rx_lock, irqflag);
 	data = fifo_get(&rxq, 1);
@@ -44,7 +44,7 @@ static size_t usart_write_int(int id, void *buf, size_t size)
 {
 	char c = *(char *)buf;
 
-	unsigned long irqflag;
+	unsigned int irqflag;
 	int err;
 
 	spin_lock_irqsave(tx_lock, irqflag);
@@ -70,7 +70,7 @@ static size_t usart_write_polling(int id, void *buf, size_t size)
 static void isr_usart()
 {
 	int c;
-	unsigned long irqflag;
+	unsigned int irqflag;
 
 	if (__usart_check_rx()) {
 		c = __usart_getc();
@@ -159,7 +159,7 @@ static int kbhit()
 
 static void fflush()
 {
-	unsigned long irqflag;
+	unsigned int irqflag;
 
 	spin_lock_irqsave(rx_lock, irqflag);
 	fifo_flush(&rxq);
