+++ include/kernel/buddy.h
@@ -8,23 +8,23 @@
 
 struct buddy_freelist_t {
 	struct list_t list;
-	unsigned long *bitmap;
+	unsigned int *bitmap;
 };
 
 struct buddypool_t {
 	struct buddy_freelist_t free[BUDDY_MAX_ORDER];
 
-	unsigned long nr_free;
-	unsigned long nr_pages;
+	unsigned int nr_free;
+	unsigned int nr_pages;
 
 	spinlock_t lock;
 };
 
 #include <kernel/page.h>
 
-struct page_t *alloc_pages(struct buddypool_t *pool, unsigned long order);
+struct page_t *alloc_pages(struct buddypool_t *pool, unsigned int order);
 void free_pages(struct buddypool_t *pool, struct page_t *page);
-void buddy_init(struct buddypool_t *pool, unsigned long nr_pages,
+void buddy_init(struct buddypool_t *pool, unsigned int nr_pages,
 		struct page_t *array);
 
 static inline unsigned int log2(unsigned int v)
